+++ ving/docs/change-log.md
@@ -7,6 +7,7 @@ outline: deep
 
 ### 2024-05-22
 * Fixed: leaking id in robot avatar #153
+* Fixed hydration mismatch on copy to clipboard button.
 
 ### 2024-05-14
 * Removed extraneous import from rest endpoint.